rename log to graph
